SlangBrowser - an addition to VMMaker for interactive browsing of C source for the interpreter and plugins.
An additional change set is required to enable browser options, see http://wiki.squeak.org/squeak/5916

Implements #asCString and #asCString: for object memory, interpreter, and plugin classes. Adds browser support for all Browser and MessageSet browsers.

In a browser, the 'What to show' menu includes selections for 'translate to C' and 'translate to inlined C'.

A cache of code generators is maintained in VMMaker to provide fast display of generated C source in browser windows. If a translatable class or method is changed, all affected cache entries are invalidated, and new code generators are created on demand.