pushed up the disassemble methods from BochsIA32ProcessorAlien to CogProcessorAlien. For that to work, I also changed the specifict register names to their abstract counterpart.
Started implementing the ProcessorTrapSignal way of telling the simulator what kind of error occured in the emulator. There might be a huge problem down the road, because the ARM simulator traps branches and jumps, only after the pc has been changed. For calls, this is no problem, because we have the lr. But what about long jumps?

Removed the opcode-methods, which were only used in filling the OpcodeTable for ProcessorTraps. This way of identifying the last instruction was not applicable, because some of those opcodes are implemented in several ways, some of which require more than one ARM instruction. For now, see #instructionTypeAt:in:.