CogVM source as per VMMaker.oscog-eem.225.
Fixes for cog issue 109, base frames and CallPrimitive, and merge
with VMMaker.oscog-lw.224.

Cogit:
StackToRegisterMappingCogit, cog issue 109.
Fix pc mapping for popped folded constants as in, e.g. 1-1.  Need to
check for annotateUse on popping a stack descriptor.

CoInterpreter:
Fix makeBaseFrame: for methods with CallPrimitive that get restarted.
e.g. on:do:.  We could make the CallPrimitive bytecode check for
being at the start of a method, but I have decided, for strictness, to
make executing CallPrimitive an error for the moment.  This means
that thater we could use it to embed primitive calls in the middle of methods.

Streamline once again activation sequence to make setting method
(actually bytecodeSetSelector) faster and on primitive failure to
increment pc past CallPrimitive before checking for err code store.

Make ensureMethodIsCogged: answer the cogged method, again for efficiency.

Add tracing of stack overflows.

Fix bug in printing of bytecode addresses in long/printOop: on CompiledMethods.

Fix bug in printStringOf: so that it prints ... when truncating.

Add missing case to isNullExternalPrimitiveCall:.

Streamline long/printReferencesTo:.

Slang:
Improve formatting of code for cppIf:ifTrue:.


git-svn-id: http://squeakvm.org/svn/squeak/branches/Cog@2624 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: 816daecdec7a9bed22a1fe7dc094c9580a8fbd1a