[New[Co]]ObjectMemory: Fix freeing of objects for becomeForward:.  Remove freed young roots from the rootsTable.  Filter freed objects pointet to from the extraRootsTable (because these locations can change it is wrong to remove entries from the extraRootsTable).
Abstract away use of the RootBit into isYoungRoot[Header]:

StackToRegisterMappingCogit:
Fix marshalling of absent receiver sends.  The items beneath the
arguments (and to-be-pushed receiver) must be spilled before
the receiver is pushed.

Improve code quality for numArgs > numRegArgs sends when
receiver is not a spill and there are no uses of ReceiverResultReg
amongst args.  e.g. now avoids loading ReceiverResultReg from stack
in code such as 1 with: 2 with: 3 .