VMMaker.oscog-eem.261.mcz Merge with VMMaker.oscog-lw.260, but put push of link register where I think it should go in trampolines.
Move determination of the ammount of headroom to the platform in
osCogStackPageHeadroom (in the various sqFooMain.c files).

Provide a routine to monitor the ammount of unused headroom,
which requires the stack memory be zeroed before use.  Assume
the platform will provide a -reportheadroom flag for enabling the
report.  Provide primitiveMinimumUnusedHeadroom for in-image
access.

Add some asserts to check that a page's frame pointer is always
in range (setHeadFP:andSP:inPage: already did this).

Implement some missing prototypes in InterpreterProxy.  Add a
comment to preDeclareInterpreterProxyOn: to eyeball all prototypes.