CogVM source as per VMMaker.oscog-eem.261.
Move determination of the ammount of headroom to the platform in
osCogStackPageHeadroom (in the various sqFooMain.c files).

Provide a routine to monitor the ammount of unused headroom,
which requires the stack memory be zeroed before use.  Assume
the platform will provide a -reportheadroom flag for enabling the
report.  Provide primitiveMinimumUnusedHeadroom for in-image
access.

Add some asserts to check that a page's frame pointer is always
in range (setHeadFP:andSP:inPage: already did this).

Add the GdbARMPlugin.

Correct cygwin HowToBuilds.

Nuke the sparseimage in the Newspeak mac installer.

Add the named serial port read & write primitives to the SerialPlugin.

Fix change history upload in uploadvms script.


git-svn-id: http://squeakvm.org/svn/squeak/branches/Cog@2676 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: dccdcdcbcd3ce98b2e1d8e07fb89ff4e4cb49b2e