Sources as per VMMaker-dtl.315
VMMaker 4.11.5

Remove PixelValuePeekPlugin from defaultUnixSpec.

New SlangTest tests for loop limit code generation.

From VMMaker.oscog-eem.279 and VMMaker.oscog-eem.282:

Fix translation of to:by:do: loops so that the limit is not re-evaluated
on each iteration if it may have side-effects.
As part of this change make TMethod locals a Set, and sort only when
emitting locals.  Alas this causes a number of methods to change.

Fix SmartSyntaxPlugin support method affected by the fixes
for generating to:by:do: loops in VMMaker.oscog-eem.279.



git-svn-id: http://squeakvm.org/svn/squeak/trunk@2716 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: c7a34cfca4f3f328d1701613b9f241d7f12ecac3