VMMaker 4.12.5 VMMaker-tpr.321
Add the VMMaker parts to support a new faster BitBLT plugin.
In copyBIts we can now branch off to a specialised function that will
make use of more tightly code routines for many important BLT rules and
cases. For ARM cpu platforms there are also very specialised assembler
routines for some cases - but many are in C and suitable for any
platform.


git-svn-id: http://squeakvm.org/svn/squeak/trunk@2745 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: 9e418c8e6243659da32752b727c43dfb8b14ec3a