Fix bug in transferTo:(from:) when doing a code compaction when ensuring there is a machine code method when switching to a process whose context has a machine code pc.
Add an assert to commenceCogCompiledCodeCompaction to catch
the actual bug (pushing the instructionPointer twice).

Improve inlining via inlineSend:directReturn:exitVar:in: by refactoring
argAssignmentsFor:args:in:'s innards.  Now global variables are
inlined if they are only read within the code being inlined.

Implement warningat in term of warning so one only has to remember
to set a breakpoint in warning, not both.