CogVM source as per VMMaker.oscog-eem.314
Fix bug in transferTo:(from:) when doing a code compaction when
ensuring there is a machine code method when switching to a
process whose context has a machine code pc.

Add an assert to commenceCogCompiledCodeCompaction to catch
the actual bug (pushing the instructionPointer twice).

Improve inlining via inlineSend:directReturn:exitVar:in: by refactoring
argAssignmentsFor:args:in:'s innards.  Now global variables are
inlined if they are only read within the code being inlined.

Implement warningat in term of warning so one only has to remember
to set a breakpoint in warning, not both.


git-svn-id: http://squeakvm.org/svn/squeak/branches/Cog@2760 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: 3e4b613d75f9fb298468e5984cce115200863a5c