VMMaker.oscog-eem.331 Speed-up Newspeak significantly (e.g. -28% in one compile-intensive benchmark) by maintaining unpaired methods (compilations of anonymous accessors) on a linked list instead of searching the entire method zone.
Do this by adding a NewspeakCogMethod and surrogates that add a nextMethod
field.  Update surrogate accessor generation so that accessors can answer and
receive surrogates and nils.  Also add a nop before the dynSuperEntry so as to
change its alignment now that (in the Newspeak VM) CogMethod has changed size
in gaining a field.

Send fetchClassOf: to the right receiver in
explicitOuterReceiver:withObject:withMixin:

Slang: fix various bugs in header and typedef generation so that only one CogMethod
typedef is produced.