CogVM source as per VMMaker.oscog-eem.333
Speed-up Newspeak significantly (e.g. -28% in one compile-intensive benchmark)
by maintaining unpaired methods (compilations of anonymous accessors) on a
linked list instead of searching the entire method zone.

Do this by adding a NewspeakCogMethod and surrogates that add a nextMethod
field.  Update surrogate accessor generation so that accessors can answer and
receive surrogates and nils.  Also add a nop before the dynSuperEntry so as to
change its alignment now that (in the Newspeak VM) CogMethod has changed size
in gaining a field.

Eliminate CogStackPage, collapsing it onto InterpreterStackPage
(it only has a class side).

Use temporaryCountOfMethodHader: instead of tempCountOf: in
SimpleStackBasedCogit>>compileFrameBuild (header is in hand).


git-svn-id: http://squeakvm.org/svn/squeak/branches/Cog@2776 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: 1b3279e206a60558bb6634a3edcef6223928edb4