Commit a barely functional markAndTrace: using a mark stack.
Introduce an obj stack datatype, a stack composed of pages that
grows but does not release memory back on shrink (when should it
do that?).  Use this for the mark stack and the ephemeron queue.

Add SpurContiguousObjStack and use it for the unscanned
ephemerons during global and/or incremental  gc.

Re-categorize the GC methods to group global, scavenging and
incremental GC together.