Rethink the sorting of free space for exact-fit compact post global GC.  Since the sweep to free unmarked objects is linear through old space it can also sort free space.  So freeUnmarkedObjectsAndNilUnmarkedWeaklingSlots => freeUnmarkedObjectsNilUnmarkedWeaklingSlotsAndSortAndCoallesceFreeSpace (glup ;-) ).
Make sure that bridges are marked so they won't be freed in
fUONUWASACFS.

Add freeChunkPrevIndex to link next nodes to their precedessors
so that free chunks can be quickly removed during coallescing.

Implement an iterative freeTreeNodesDo: so that Slang can inline the
block arguments to it.

Refactor the code that removes a tree node in
allocateOldSpaceChunkOfBytes: et al, out into
unlinkSolitaryFreeTreeNode:

Still have to reverse small list heads post-sweep and account for free space properly.