VMMaker.oscog-eem.485.mcz Fix ordering of removal of final return and recording of declarations. i.e. move them from TMethod>>inferReturnTypeIn: to CCodeGenerator>>inferTypesForImplicitlyTypedVariablesAndMethods. Don't do them more than once.  Pass in code generator to both to allow logging of errors.  Check that recorded declarations are for extant variables.  Correct a few methods in which this wasn't true.
Refactor the return typer determination into addTypesFor:to:in:
so it can recurse.  Add support for some arithmetic ops.

Add asVoidPointer convenience and use it in several mem:cp:y:/
mem:mo:ve: contexts, as welkl as to replace clumsier
cCoerceSimple:'s.

Force the type of all entries in the primitive table to be void in
StackInterpreter class>>preGenerationHook:.