VMMaker.oscog-eem.501.mcz Refactor (pop)storeLiteralVariable in the StackInterpreter so that all such bytecodes use storeLiteralVariable:withValue:.  Put breakpoints in the simulator's pushLiteralVariable: & storeLiteralVariable:withValue: to catch forwarded objects.  This for debugging restarting prims that have failed because of forwarded args. Use ^self in the clients (and in itemporary:in:put:) now that Slang can cope.
Spur:
Revise (and refactor) markAndTrace: so that
- pure bits objects are not pushed on the stack
- classes of objects are traced
- checking for immediates, pure bits and weakness is done in markAndShouldScan:
- class processing uses the classTableBitmap to expunge duplicate entries.

Remove the bogus postCompactScanClassTable call from
globalGarbageCollect.

Add followForwardingPointersInScheduler and use it post-become.

Refactor the all*Object*Do: methods to use isEnumerableObject:.
Change the classIsItselfPun so that such objects answer true to
isEnumerableObject:.

Make sure that allocateSlots:format:classIndex: is inlined and that
allocateSlotsInOldSpace:bytes:format:classIndex: and
eliminateAndFreeForwarders are not.

Provide class side utilities to show the class puns and the formats in order.

Add the missing longPrintReferencesTo:.

Fix printOopShortInner: for simulation.

Mimic the free space determination in sqMacMain.c in
the simulator's openOn:. [N.B. this fails to mimic because of
bugs in or around readImageFormatFromFile:StartingAt:.

Slang:
Fix generation of asserts within struct classes in
shouldExcludeReceiverAsFirstArgument:.