CogVM source as per VMMaker.oscog-eem.517
Move the defines for BaseHeaderSize BytesPerWord BytesPerOop out of interpreter,
cogit or plugin source and into interp.h, and fix generateCPtrAsOop:on:indent:
to print BaseHeaderSize symbolically so that plugin source can be safely shared.

Include the executable name in the crash.dmp report.

Fix uninitialization of newTargetMethodOrNil & errorSelectorOrNil
in ceCPICMiss:receiver:.

Do a better job of dead code removal, accepting constant comparisons.

Lots of changes to Spur, find in Monticello but not relevant here as
Spur is not yet ready for general consumption.
Add the Cog Spur source to accompany the build dirs (doh!).

Internal:
Rewrite markLiteralsAndUnlinkIfUnmarkedSend:pc:method: &
markLiteralsAndUnlinkIfUnmarkedSendOrPushImplicit:pc:method:
so that markAndTraceLiteral: is only invoked with literal args and
cacheTagIsMarked: is only applied to cache tags.  This allows Spur
to answer true for cacheTagIsMarked:.

Refactor (pop)storeLiteralVariable in the StackInterpreter so that all
such bytecodes use storeLiteralVariable:withValue:.

JFK RIP.


git-svn-id: http://squeakvm.org/svn/squeak/branches/Cog@2817 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: 58636e997b16f85d72b34328fd0fc82a6678d216