VMMaker.oscog-eem.519.mcz Commit a costly mistake before undoing the damage.  If the code zone sits between newSpace and oldSpace then isMachineCodeFrame: et al need two boundary checks, > newSpaceLimit & < oldSpaceStart. If the code zone sits beneath newSpace then copyAndForward: needs two boundary checks, < newSpaceLimit & > newSpaceStart.  So IIABDFI implies reverting to the code zone being beneath newSpace, and fixing copyAndForward:.  Tube.  3 days. Ouch.
Change Spur's primitiveFullGC to answer the byte size of the largest
free chunk, which is what is needed to know whether a new segment is
needed.  Add primitiveSizeInBytes[OfInstance] so the image doesn't
have to assume anything about header sizes.

Fix checkMaybeObjRefAt: to filter-out the 0's from unused method
parameters in closed PICs.

Fix a return-type conflict in printOopShortInner:.

Slang:
Fix Integer << Integer to cast to 64-bits if the shift overflows 32-bits.