Fix translation of var := TStmtList so that the value of the last statement is assigned, not the first.
Fix the concatenation of the assert-failure reason in
isValidObjStackPage:myIndex:firstPage: so that it works in C.

Remove the assert in updateRootOfObjStackAt:with: so that the
intermediate state of a partially followed objStack is not reported
as an assert-failure.

Add a print routine for pages of an objStack.

[ All to fix broken assert code :-) ]

Slang:
Generate the CoInterpreter *before* the Cogit so that the
objectMemory and interpreter classes have initialized *before* the
Cogit is generated.  Otherwise the Cogit can be generated with values
from the wrong interpreter/objectMemory.