CogVM source as per VMMaker.oscog-eem.545.
Fix setting of endOfMemory on load in SqueakV3 VMs.

Spur:

Slight streamlining in followForwardingPointersInStackZone:, plus fixingfollowjng of frameMethodObject vs frameMethod in machine-code frames.

Fix the concatenation of the assert-failure reason in
isValidObjStackPage:myIndex:firstPage: so that it works in C.

Slang:
Generate the CoInterpreter *before* the Cogit so that the objectMemory andinterpreter classes have initialized *before* the Cogit is generated.  Otherwisethe Cogit can be generated with values from the wrong interpreter/objectMemory.


git-svn-id: http://squeakvm.org/svn/squeak/branches/Cog@2820 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: 471e58f2cd3c5f115f3f97f5188e0814177daa9a