Fix allocateSlots:format:classIndex: to not chew into the interpreter reserve.  That's only for allocateNewSpaceSlots:format:classIndex:.
Fix enumeration of newSpace objects to be in addfressSpace order,
survivor space first, eden second. This allows printObjectsFrom:to:
to determine the start of a newSpace object correctly.

Fix the store check on SpurMemMgr>>clone:.

Add cmUsesMethodClass flag and use it to reduce
 use of followForwardedLiteralsIn: in CoInterpreter's
followNecessaryForwardingInMethod:.
Make makeBaseFrameFor: call followNecessaryForwardingInMethod:.

Add StackToRegisterMapping2RegArgsCogit although it is decidedly
non-functional at the moment.