Sources per VMMaker 4.13.4 (VMMaker-dtl.344)
Fix slang inlining regression introduced in VMMaker-dtl.339 caused by
TMethod>>tryToInlineMethodsIn: that excluded nodes with selector #cCode:inSmalltalk:
which must remain inlineable to support various methods for 32/64 bit image
format from common code base, as well as for low level MemoryAccess in slang
that must be fully inlined for performance.

Change FloatArrayPlugin>>primitiveDivFloatArray and BitBltSimulation>>primitiveDisplayString
to send intAtPointer and byteAtPointer to self rather than to interpreterProxy,
because they are not defined in the interpreter proxy and cause compilation
errors when slang inlining is turned off. These should be added to the interpreter
proxy (or arguably to a hypothetical objects memory proxy) but send to self is
a lesser evil for now. 

These changes restore a very small performance gain for the standard VM
due to the inlining fix, and also restore inlining of low level MemoryAccess
methods when added to the VM build. Inlined MemoryAccess methods are essentially
equivalent to the C macros in sqMemoryAccess.h but can be run in simulation,
and play nicely with C debuggers and profilers.


git-svn-id: http://squeakvm.org/svn/squeak/trunk@2878 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: d05bcee62f834a46b4c20948c4ae1e01b2ff2c99