VMMaker.oscog-eem.642.mcz Mostly simulator work. Rationalize signed[32|64]BitIntegerFor:. Fix odd bytes calculation in Spur>>allocateBytes:classIndex:.
Refactor the Spur leak checker so that references to garbage- collected classes are tolerated mid-way through GC but not before
or after.

Make primitiveSine etc simulate correctly when given Float nan as
operand.

Make asPositiveIntegerObj simulate correctly.

Fix screen update when large simulator windows are opened.

Fix screen update in BitBltSimulation>>unlockSurfaces and delete
the bogus fullDisplayUpdate on every 1000 bytecodes.

Fix access to the cmShiftTable in BitBltSimulation for negative shifts.
Fix the unitSize asserts in loadColorMap for simulation.
Fix simulation of primitivePixelValueAt[X:y:]

Increment byteCount by microseconds in ioRelinquishProcessorForMicroseconds:
so that microsecond clock advances (but this should be improved
by adding a separate microsecond clock).

Rewrite checkIfValidObjectRefAndTarget:pc:cogMethod: to use the
new accessors, but the assert for a cacheTag is still wrong for Spur.

Speed up StackInterpreterSimulator by adding a flag to control the
use of assertValidExecutionPointers on each bytecode, with a toggle
in the utilities menu.