CogVM source as per VMMaker.oscog-eem.656
Fix store check call from machine code which in Spur was showing that ClassReg
(%ecx, a caller-saved reg) was not being saved when calling ceStoreCheck:.
This /could/ be the cause of crashes in the clang build (or at least one major
cause).  It didn't show in the classic VM because the store check is simpler.

Provide __stringify[Num] even when NDEBUG=1 for manual hacking in of warnings.

Fix revertIfEssentiallyUnchanged for OSProcess plugins.

A few plugins have been regenerated with the new array literal index prefixes.

Spur:
Ugh, what a bug hunt.  Move the expungeDuplicateAndUnmarkedClasses: send from
markObjects to globalGarbageCollect where it belongs and hence prevent
allObjects and allInstancesOf from breaking the class table.

Rationalize all but one of the various sends of runLeakCheckerForFullGC:exclude-
UnmarkedNewSpaceObjs:classIndicesShouldBeValid: to runLeakCheckerForFullGC:

Streamline copyAndForward: and forward[Survivor]:to: and clients
of setFormatOf:to: & setClassIndexOf:to:.

Fix copyAndForward: to add tenured objects to weak and ephemeron lists if
required.  Because the weak and ephemeron lists are threaded through forwarding
corpses they can indeed refer to old objects, and just cuz a weak/ewphemeron
object is tenured doesn't mean it doesn't need to be scanned post-scavenge.

Relax the leak-check in Spur fullGC to allow invalid class indices at any point
except after fullGC, at which point the classTable should have been purged.


git-svn-id: http://squeakvm.org/svn/squeak/branches/Cog@2883 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: 353036ab5ea18d24a05c638266185260ab820af0