VMMaker.oscog-eem.661.mcz Sista: Reload TempReg from ReceiverResultReg on return from counter trip, so that branch retry tests the right value. Add a primitive to reset branch counters in a method.
Spur: abstract out the average object size calculation into its own
method (it'll be needed for pig compact).

Simulator:
Fix sqFile:Read:Into:At: for stdout when stdoiut is the Transcript
(Pharo startup indirectly tries to set line-end convention of stdout).

Implement remapping in remapOop:in: (simulating a Sista image
resulted in a GC in one of the LargeInteger routines).