CogVM source as per VMMaker.oscog-eem.663
Fix pc-mapping for NewspeakV4.  Dynamic super sends should
/not/ be annotated with IsNSSendCall, but wth the vanilla IsSendCall.
This fixes a bug converting an interpreter activation of a method
with a loop and a duynamic super send to a machine code frame.

Sista:
Add sistasrc & sistamacbuild.  It's time ;-).
Reload TempReg from ReceiverResultReg on return from
counter trip, so that branch retry tests the right value.
Add a primitive to reset branch counters in a method.


git-svn-id: http://squeakvm.org/svn/squeak/branches/Cog@2887 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: 5029f9a00dc17780b654dfc5d254a816f7f67a48