VMMaker.oscog-eem.665.mcz Sista: Not understanding what I'm doing has caused confusion and delay.
The return from ceCounterTripped: should indeed jump to the retry
of the first counter decrement.  But the result comes in through
cResultRegister, which may be the same as TempReg which the code
uses to hold the boolean tbeing tested.  So TempReg should /not/ be
reloaded with ReceiverResultReg.

Remember to count the untaken branch in inlined special selector comparisons.

Simulator:
Change the prompt for the break block to present a template.