CogVM source as per VMMaker.oscog-eem.665
Sista:
Not understanding what I'm doing has caused confusion and delay.

The return from ceCounterTripped: should indeed jump to the retry of the first
counter decrement.  But the result comes in through cResultRegister, which may
be the same as TempReg which the code uses to hold the boolean tbeing tested.
So TempReg should /not/ be reloaded with ReceiverResultReg.

Remember to count the untaken branch in inlined special selector comparisons.


git-svn-id: http://squeakvm.org/svn/squeak/branches/Cog@2888 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: b5af95c08803194a0c6b49cb664344ed1662c06e