VMMaker.oscog-eem.666.mcz Sista: Fix several bugs in the primitiveSistaMethodPICAndCounterData primitive.  Fix consistency between monomorphic and polymorphic send tuples, always answering pc, class, method/selector. Fix collecting data from PICs with MNU cases.  Remove a stray halt. Make sure primitive fail points return. not continue on.
Eliminate dead conditional branch elimination so that the send and
branch data collection is not confused by branches for which no
counting code was produced.

Cogit: since the annotation arg is no longer used, refactor
targetMethodAndSendTableFor:annotation:into: to
targetMethodAndSendTableFor:into:.

Simulator:
Add missing asVoidPointer on methods.
Add instantiateClass:indexableSize: for the facade.
Provide a cog: entry point for in-image compiling a method.