VMMaker.oscog-eem.670 CoInterpreter: Garbage collect/remap the primTraceLog correctly.  If a GC happens ver early in start-up the log circular buffer may not be full and the existing code assumed it always was.
Cogit:
Fix an assert fail in mapFor:bcpc:performUntil:arg: (this for
primitiveClass where the class table reference can be the first
map entry).

Newspeak:
Fix a bug in remapIfObjectRef:pc:hasYoung: with dynamic super
sends which could compute an invalid target method.

Rename pushExplicitOuterSendReceiverBytecode et al to
pushExplicitOuterReceiverBytecode et al.  These are not sends.

Sista:
Fix picDataFor: which must return integers or oops, never nil.
Cope with doubleExtendedDoAnythingBytecodes which /may/ be
sends or not, and hence are problematic for gathering send data.
Generate correct counting code for #==, which implies it can't be
frameless in Sista.
Annotate the right address (the return address of the
ceMustBeBooleanFoo call) in counting genJump:if:.
Sista now starts up a Pharo image and can gather send and branch
data for all methods.

Simulator:
Evaluate the cogit's break block on each bytecode in the
CogVMSimulator.
Add missing ObjectMemory>>ioLoadModule:OfLength:.
Cope more gracefully with not finding an image.