CogVM source as per VMMaker.oscog-eem.670
CoInterpreter:
Garbage collect/remap the primTraceLog correctly.  If a GC happens
very early in start-up the log circular buffer may not be full and the
existing code assumed it always was.

Cogit:
Fix bug in unlinkSendsOf:isMNUSelector: (primitiveFlushCacheBySelector)
where old code could free the method of an active frame.

Fix an assert fail in mapFor:bcpc:performUntil:arg: (this for
primitiveClass where the class table reference can be the first map entry).

Newspeak:
Fix a bug in remapIfObjectRef:pc:hasYoung: with dynamic super
sends which could compute an invalid target method.

Rename pushExplicitOuterSendReceiverBytecode et al to
pushExplicitOuterReceiverBytecode et al.  These are not sends.

Sista:
Cope with doubleExtendedDoAnythingBytecodes which /may/ be
sends or not, and hence are problematic for gathering send data.
Generate correct counting code for #==, which implies it can't be
frameless in Sista.
Annotate the right address (the return address of the
ceMustBeBooleanFoo call) in counting genJump:if:.
Add Spur support for Sista (shorten:toIndexableSize: et al).
Implement instantiation in picDataFor:into: in terms of
eeInstantiateClassIndex:format:numSlots:.



git-svn-id: http://squeakvm.org/svn/squeak/branches/Cog@2890 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: 35d773992ebeb9243d120b57e8bc984cbba4abd2