CogVM source as per VMMaker.oscog-djm.688
ARM32FFIPlugin:
Change dummyFloatFunction to loadFloatRegs since it is a better name. Define this
function elsewhere (platforms/Cross/plugins/SqueakFFIPrims/sqFFIPlugin.c) so that
the C optimizer cannot optimize it away. This fixes a problem when calling out to
functions taking float parameters when compiling with optimization level > 0.

VMMaker image:
Add an expression to the workspace to generate plugins.


git-svn-id: http://squeakvm.org/svn/squeak/branches/Cog@2908 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: e742ba7e4e948e49ed4aeb04cb786c6ffac4cc7f