VMMaker.oscog-eem.689.mcz Spur: Implement "pig compact", a much more functional compaction algorithm that works by doubly-linking free chunks in address order, therefore allowing e.g. easy enumeration of the objects between the penultimate and ultimate free chunks.  Hence the algorithm moves all the objects it can at the end of memory to free chunks at lower addresses.  It is piggish for several reasons: 1. it is greedy, using parts of a free chunk, not looking for a best or perfect fit. 2. it is greedy trying to move a run of objects at a time 3. it deals with large objects ("pigs") by searching the free list for a free chunk large enough to hold the pig.  (and what constitutes a pig remains to be tuned; currently it is 8 * the average object size.
Write the totalFreeOldSpace to the image header immediately
following the size of the first segment.  This to allow better
determination of how much free space to allocate on startup.

Use cCode:inSmalltalk: to encapsulate the printing of the gc
stages, and delete the overrides in the simulator subclasses.

Make sure the last bridge is marked to simplify
coallesceFreeChunk:

Rename averageObjectSize to averageObjectSizeInBytes.

Add printOopsFrom:to:


General VM:
Nuke unused writeImageFile:

Fix a slip and a bug in printObjectsFrom:to:

Add a convenience to open a multi-window browser on the
Spur classes, and relevant allies.


Cogit:
Add inline pragmas to the concretize methods that were lacking them.