VMMaker.oscog-eem.693.mcz Eliminate the abuse of prepareToBeAddedToCodeGenerator: to remove superclass methods.  Make method conflict checking use shouldIncludeMethodFor:selector: asnd allow option: pragma methods to override methods in other hierarchies, albeit with a warning.
Either delete or simplify a whole lot of ugly
prepareToBeAddedToCodeGenerator: hacks.

Use the option: approach to organize the numRegArgs
implementations, which allows inlining and the necessary
dead code elimination in CoInterpreter to avoid implementing
the register enilopmarts in SimpleStackBasedCogit.

Fix slip bug in ObjectMemory>>isContextHeader:

externalSetStackPageAndPointersForSuspendedContextOfProcess:
must be marked <inline> to be reliably inlined in transferTo:/transfer:to: et al.