VMMaker.oscog-eem.697.mcz Spur pig compact: Fix a bug in free list enumeration in pigCompact. sortedFreeListPairwiseReverseDo: must cope with the case where moving s run of objects alters the free list.
Fix a bug in moveARunOfObjectsStartingAt:upTo: which
would not move the first object if it was the only object
in a run.

Fix a bug in freeUnmarkedObjectsAndSortAndCoalesceFreeSpaceForPigCompact
which could leave a free object with a 0 slot count, which
falls foul of the object enumeration invariant that a header
word must be all zeros (for catching wild missteps).

Fix a bug in sweepToCoallesceFreeSpaceAndRebuildFreeListsForPigCompact
which would cause it to miss lots of forwarders.

VMMaker:
Update generateAllSpurConfigurations to include Spur Cog Sista.