CogVM source as per VMMaker.oscog-eem.701
Spur:
Fix bug in generated C for pigCompact.  If "free" is unsigned
then "free - self allocationUnit" is large when free is zero, so
move subtraction to addition on other side of the comparison.

Fix a mis-write/typo bug in moveARunOfObjectsStartingAt:upTo:.

Make sure all pointer comparisons in pigCompact are unsigned.

Add forwardUnchecked:to: and use it to avoid the unnecessary
store check in forwarding during compaction.

Correct the return type of bytesInObject: to sqInt.

Looking good.  Snapshots are small at last.

Slang:
Fix indenting in while loops and sequential and and or exprs.


git-svn-id: http://squeakvm.org/svn/squeak/branches/Cog@2913 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: bf27d38fa7057e715797f5f173ea6215b8c92a15