VMMaker.oscog-eem.733.mcz Use unsigned comparisons when testing if there's sufficient heap space in readImageFromFile:HeapSize:StartingAt:.
Remove break in CogVMSimulator>>ceSend:super:to:numArgs:.
Monticello should really check for break methods on commit.

Refactor cloneSimulation into cloneSImulationWindow and
cloneSimulation.  Remember the clone and the parent.  Use
this to make the Spur GC take a clone and push it off the
cliff every global GC to allow repeating GCs that hit bugs.

Add an assert to check that all weaklinks are accessible
(but the bug has already been found with the nextCorpse
encoding bug).