CogVM source as per VMMaker.oscog-eem.746
Implement a peephole in the Spur Cogit for an indirection
vector initialized with a single value  Avoid initializing the
slot in the array to nil and instead initialize it with the value.

Refactor setting byte1, byte2 & byte3 into loadSubsequentBytesForDescriptor:at:
for the peephole tryCollapseTempVectorInitializationOfSize:.

No longer inline CoInterpreter>>pre/postGCAction: for VM profiling.

Increase the number of trampoline table slots.


git-svn-id: http://squeakvm.org/svn/squeak/branches/Cog@2938 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: e3a040202d3c7997f19d861aeaa353271d0e2e3d