Spur: Implement unforwarding in inlined machine code #==. hence rewrite the /horrible/ StackToRegisterMappingCogit >>genSpecialSelectorEqualsEquals.
Unforward method literals in cog:selector: to ensure that all
examinations of constants in the Cogit are of the actual objects.

Have StackToRegisterMappingCogit>>genPushReceiverBytecode
use ReceiverResultReg if it contains self.

Speed up isForwarded: test by using the fact that the
isForwarded class pun is a power of two.

Fix a slip in an assert in unlinkSolitaryFreeTreeNode:.

Comment out the checkFreeSpace check in copyToOldSpace:bytes:format:.

Make the assert check in Spur32BitCoMemoryManager>>
freeStart: more lenient to avoid bogus assert fails from
machine code allocations.

Simulator:
Bring in-image facade for spur up to date.