VMMaker.oscog-eem.762 Move BICCqR from CogRTLOpcodes to CogARMCompiler. It's currently ARM-specific and there's a mechanism for handling processor-specific opcodes.
Revert the ceCheckForInterrupts hack.  The pre-hack code
is correct.

Streamline the genPrimReturnEnterCogCodeEnilopmart:

But most importanty, fix CogRTLOpcodes class>>initialize to
automatically initialize the CogAbstractCompiler subclasses
after initializing its opcodes, so that the processor-specific
opcodes are correctly initialized.  The simulator breaks in
a visually interesting way (read most things continue to work)
if not.