CMakeVMMakerSqueak-tty.54.mcz Linux64x86w32BitSqueakCogV3SlackwareNoGLConfig generates successfully.
The configH method is customized to support compiling the UUIDPlugin. (which subsequently fails to dynamically load, but at least the ldUUIDPlugin.so file is created. small steps. small steps)

Next up. implement setting internal state of this configuration for various build types:

#build   <---done
#buildAssert 
#buildAssertITimerHeartbeat 
#buildDebug 
#buildDebugITimerHeartbeat 
#buildDebugMultiThreaded 
#buildITimerHeartbeat 
#buildMultiThreaded 
#buildMultiThreadedAssert 
#buildMultiThreadedDebug 
#buildNone


With that infrastructure in place, all the (well, most) tools are in place to chug through all the configurations.