CMakeVMMakerSqueak-tty.56 Put in skeleton methods for a configuration to set its internal state by buildType. These methods mirror each message send in CMakeVMGenerator>generate method
setGlobalOptions
setGlobalOptionsAfterDetermineSystem
setupDirectories
preferredIncludes
includeDirectories
standardIncludes
compilerFlags
extraVMSettings
internalPlugins
externalPlugins
thirdpartyLibs
setExtraTargetProperties
postBuildActions

with a suffix of xyzBuildType. From there, on message to configure for a buildType, we access these to set those internal parameters.

Stubbed out a Step by Step Guide on setting up a configurations. Since I will be doing that for the next week, I should have that documented.

Changed and re-organized the protocols for cmake in TPlatformConfSqueak. 

added blurb in Help Documentation to that effect.