CMakeVMMakerSqueak-tty.78.mcz Linux32x86SqueakCogV3NoGLConfig  is ugly, but it compiles. Next up...replace the compilerFlags/cFlags with cC, cXX,lDFlags and libs such that the process is intuitive, repeatable and correct.
Then, flesh out the various buildTypes build.assert, etc so that they compile exactly as the GNU system does.
coreSources will change based on that, so put in infrastructure to support that.

Then, figure out how to run 'configure' and pipe in the results of config.h used by the configuration.