VMMaker.oscog-eem.795.mcz Spur: Implement forwarder following on primitive failure for side- ways calls from mahcine code.
Fix LargeIntegersPlugin>>isNormalized: for forwarders, no
longer assuming that if its arg isn't a SmallInteger it must be
a large integer.  Squash an assert fail in lengthOf:format: on
forwarders by using numSlotsOfAny:.

Make sure a forwarder has an accurate slot count, bumping it
to 1 if it was zero.

Simulator:
Fix Spur simulation of SmartSyntaxInterpreterPlugins with no
explicit simulator subclass.