CogVM source as per VMMaker.oscog-eem.794
Fix the shift for divide issues in the LargeIntegersPlugin.  Add code to
generateDivide:on:indent: to spit out checking asserts if required.

Change the SmartSyntaxPluginCodeGenerator to generate
code that ifdefs out the remapOop:in: rigmarole on Spur.

Fix inline cache for Characters in Spur.

Update mksistaarchives for new build structure on Mac (only).


git-svn-id: http://squeakvm.org/svn/squeak/branches/Cog@3026 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: d942eb5ac4d467a7da03f8488e10f516cbd2bbb4