CogVM source as per VMMaker.oscog-eem.796
Spur:
Implement forwarder following on primitive failure for sideways calls from
machine code.

Fix LargeIntegersPlugin>>isNormalized: for forwarders, no longer assuming that
if its arg isn't a SmallInteger it must be a large integer.  Squash an assert
fail in lengthOf:format: on forwarders by using numSlotsOfAny:.

Make sure a forwarder has an accurate slot count, bumping it
to 1 if it was zero.


git-svn-id: http://squeakvm.org/svn/squeak/branches/Cog@3028 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: 95267e0eb9bd6fa7c974d0bc59509fe4073dd2d5