CogVM source as per VMMaker.oscog-eem.804
Cogit:
Change the management of counters in Sista methods to hold the counters well
away from code.  This restores the performance of the Sista VMs on x86, where
when counters are close to code the writes to counters flush the instruction
cache and destroy performance.  In V3 use malloc to allocate and free them
(this means Sista V3 is not currently simulable).  In Spur, store them in
oldSpace in pinned objects (hence Sista Spur simulates just fine).

Split the enilopmarts into those that are used in the context of a call
(entering code as if from a call) and those used in other contexts (converting
interpreted method to machine code method in loops, returning from interpreter
method to machine code one, etc).  This fixes enilopmarts on RISCs (i.e. ARM)
where the ret pc must be popped into the LinkReg when entering machine code as
if from a call, but not when entering in other contexts (returning from inter-
preter to machine-code method, converting interpreter method to machine code in
a loop, etc).

Nuke a couple of unused variables in Cogit hierarchy.

Spur:
Fix bug in allocateSlotsForPinningInOldSpace:bytes:format:classIndex: and make
sure answered object is actually pinned.  Hence fix pinObject: & primitivePin.

Plugins:
Fix access to the characterTable in the ThreadedFFIPlugins.  Replace
characterTable at: with characterObjectOf:, and in Spur support wide characters.

Fix divide-as-shift issue in BalloonEnginePlugin.


git-svn-id: http://squeakvm.org/svn/squeak/branches/Cog@3032 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: cd1f54504a786eef09d78fb46ea06eacfac37759