VMMaker.oscog-eem.814.mcz On Spur, move implicit receiver caches ou of code and into pinned objects on the heap (similarly to Sista counters). Gets e.g. -2% speedup in compiling a Newspeak file.
Rename NewspeakMethod>>nextMethod to nextMethodOrIRCs
and use it for both tasks.  Nuke SMM>>allocatePinnedCounters:
in place of allocatePinnedSlots: and use this for both counters
and IRCs.

Add a hasIRC flag to CogBytecodeDescriptor and count implicit
receiver caches for their out-of-line allocation.

Make the option: processing in shouldIncludeMethodFor:selector:
cope with multiple option: pragmas.

Nuke the newInitializeBytecodeTableForNewspeakV3PlusClosures.
Can be resurrected later if the others ever get worked on.