VMMaker.oscog-eem.816 Spur: Fix bugs in clone:; allocateSlots:... may fail and cloning compiled methods still needs a store check.
Fix become performance issue of following possibly
becommed selectors by adding a read barrier to
lookupMethodInDictionary: et al.  This is much cheaper than
following all dictionaries in the classTable post become.
Control the policy switch with a class var in the hope that
an efficient eager solution can be found.

Add stats that count the causes of followForwardedObject-
Fields:toDepth: (used to track down the above issue).
Make optional via a class var.  Move stringForCString: from
StackInterpreter to the object memories.

Reposition ensureNoForwardedLiteralsIn: and replace
cePositive32BitIntegerFor: with positive32BitIntegerFor:.

Remove the assert check in isForwarded: to make sure it is
inined.  Mkae sure possibleRootStoreInto: is not inlined.

StackInterpreter:
Extract the vars initialized to zero or false to an initialize
method from initializeInterpreter:.

Slang:
Fix option processing so that interpreterClass and
objectMemoryClass class vars can be tested (for the
above class vars).