VMMaker.oscog-eem.818.mcz Spur: Fix bug in scanClassPostBecome:effects: with new lazy selector following policy by... throwing it all away.
The read barriers on method lookup (of the methodClass
association in super sends, of the superclass link, of
method dictionaries, method dictionary arrays, selectors
and methods is cheap.  So replace scanning classes and
method dictionaries in te class table post become with
read marriewrs on methodClass, superclass and method
dictionary etc on lookup.

The read barrier on an object from which we are going to
fetch state (such as a class or method dictionary) is
essentially free on modern machines because the class
index and the state very likely share a cache line, and the
register code for testing is so cheap compared to memory
access.  Further the read barrier on selectors is cheap
because the method lookup cache is effective in reducing
the number of message lookups and because we don't
have to check for forwarding of nil entries.

So nuke all the followNecessaryForwardingInMethod:
machinery including the cmUsesMethodClass hack.
Nuke scanClassPostBecome:effects:.

Clean up, e.g. replace followNonImmediateField:ofObject:
uses with followObjField:ofObject:.