CMakeVMMakerSqueak-tty.87 Re-wrote the config.cmake generation to use 'Cmake Templates' which are smalltalk classes that wrap CMake commands. Think Seaside-For-CMake or  C-Side (heh) where the config.cmake file is created using components that render themselves.
This approach is much cleaner and intuitive than the message based CMake code generation used in the CMakeGenerator cmake commands protocol. Also, it does not supplant it or interfere with it.

Next up is bring in the  platforms/unix/config/config.h.in  layout to replace Ian's config file.

After that is extend this approach to Plugins generation using Ian's existing PluginExternal.cmake  PluginInternal.cmake templates

Once the templating system is in place, then its on to the home stretch and get this puppy done and documented.