VMMaker.oscog-eem.820.mcz Rip out the forwardingCount: measurement code.  It causes bad performance regressions (due to failing inlines?)
Put the handling of the cloning of cogged methods in the
clone: implementations, removing it from the primitive.
Add it to the pinning clone too.

Add SpurGenerationScavenger's api methods to Spur32BitCoMemoryManager's.  Hence search for option:
variables in the class pools of the interpreterClass and
objectMemoryClass

Specialize the store check trampoline generation.
Move it down to the relevant object representations.

Move setting of isRemembered flag to true into
SpurGenerationScavenger>>remember:.  Inline
possibleRootStoreInto: (given that remember: is /not/
inlined.  Call remember directly from the
ceStoreCheckTrampoline, and hence have
remember: answer its argument.