CogVm source as per VMMaker.oscog-eem.823
Reimplement the backward count in interpreted methods, storing the count
in IFrameFlags, hence eliminating the requirement that the jump happens
consecutively N times in just that method with no backward jumps in any other
method.  The Chameneos benchmark has this pattern and so two mehtods don't
get jitted.  This eliminates lastBackwardJumpMethod & backwardJumpCount.

Add support for not inlining in gcc under optimization (for profiling).  


git-svn-id: http://squeakvm.org/svn/squeak/branches/Cog@3043 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: 7fd02e16620b26a2b6fd63eac84cd1a6428046ba