VMMaker.oscog-eem.830.mcz Revise markAndTrace: given that markAndShouldScan: is inlined within it (see Slang changes below).  Move the ephemeron processing into markAndShouldScan: out of the now unused numStringSlotsOf:ephemeronInactiveIf: circumlocution.  Add activeAndDeferredScan: and numStrongSlotsOfInephemeral: in place of the double negative inactiveOrFailedToDeferScan: and hence inline numStrongSlotsOfInephemeral:. Increase the traceImmediatelySlotLimit.  These changes plus the 2 repeats for compaction speed up global GC by at least x2.
In the wake of the inlining changes below, split
lookupInMethodCacheSel:classTag: into it and
inlineLookupInMethodCacheSel:classTag:, and use the inline
version in internalFindNewMethod.

Slang:
Add support for inlining into the condition of ifTrue:/ifFalse:
when it is marked as inline. Transform
	expr1 ifTrue:/ifFalse: [^expr2]
by inlining ^expr2 into expr1.  Transform
	expr ifTrue:/ifFalse: [statements]
by replacing ^boolean occurrences in expr with gotos.