CogVM source as per VMMaker.oscog-eem.830
In the wake of the inlining change below, split lookupInMethodCacheSel:classTag:
into it and inlineLookupInMethodCacheSel:classTag:, and use the inline version
in internalFindNewMethod.

Add vmParameter 52 to answer the capacity of the root table
(in Spur a.k.a. the rememberedSet).

Delete all the explicit initializing to nil in primitiveVMParameter
since the array is instantiated normally (filled with nil).

Spur:
Add vm parameter 53 to answer the number of segments.

Move the rememberedSet into a pinned object in oldSpace.  Allow it to grow
on demand, starting at 1k entries, doubling on each grow.  Make sure to abort
if the attempt to grow the remembered set fails.  Try to grow by another 1k
slots if doubling fails, then abort.

Revise markAndTrace: given that markAndShouldScan: is inlined within it (see
Slang changes below).  Move the ephemeron processing into markAndShouldScan:
out of the now unused numStringSlotsOf:ephemeronInactiveIf: circumlocution.
Add activeAndDeferredScan: and numStrongSlotsOfInephemeral: in place of the
double negative inactiveOrFailedToDeferScan: and hence inline
numStrongSlotsOfInephemeral:.  Increase the traceImmediatelySlotLimit.  These
changes plus the 2 repeats for compaction speed up global GC by at least x2.

Change the defaultEdenBytes to 4Mb.

Allow the number of compaction passes to vary, 2 on GC, 3 on GC for snapshot.

Slang:
Add support for inlining into the condition of ifTrue:/ifFalse:
when it is marked as inline. Transform
	expr1 ifTrue:/ifFalse: [^expr2]
by inlining ^expr2 into expr1.  Transform
	expr ifTrue:/ifFalse: [statements]
by replacing ^boolean occurrences in expr with gotos.



git-svn-id: http://squeakvm.org/svn/squeak/branches/Cog@3051 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: f90cd0af94d9adc9d3ab7520618d99f5258d4e76