CogVM source as per VMMaker.oscog-eem.831
Spur:
Fix bug with class table management and two-way become.  Because two-way become
may do an in-place swap obj1 & obj2 in SpurMemoryManager>>doBecome:and:copyHash:
may not be forwarded after the inner become.  Hence they should not be followed
if not forwarded.  The bug manifested as Object's identityHash changing:
superclass is the first slot in a class. Following an unforwarded subclass of
object yields Object.  Setting the hash bits of the followed object smashes
Object's identityHash.

Thanks to Stephane Rollandin for finding the bug.


git-svn-id: http://squeakvm.org/svn/squeak/branches/Cog@3052 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: 0ac6dff60e520aaaeec7b49f23d9da0de552630b