VMMaker.oscog-eem.883.mcz Spur: Change become to flush method caches for becommed classes.
I'm not sure this is necessary but strangely the new
code is faster than 882 for the Newspeak bootstrap.
So I'm committing this to allow me to compare the
two.

Changes are to
- add a BecameActiveClassFlag to becomeEffectsFlags
- refuse to do an in-place become on active classes.
- flush the caches before the post-become scan of
the class table so that forwarded classes can be
spotted.  i.e. if a class index in a method cache is that
of a forwarded class then it was becommed and the
cache entry can be voided.