Fix regression in youngReferrers management in V3 become introduced in VMMaker.oscog-eem.882. We /must/ priune young referrers if mapObjectReferencesInMachineCodeForBecome removes a cog method from youngReferrers because it may get added back and youngReferrers cannot contain duplicates.
Fix potential bug in Spur become argument validation.
Check for immediates needs to come /after/
following forwarders since an object can
become-forward to an immediate.

Fix cPICHasForwardedClass:; this did /not/
enumerate the classIndices in a Spur Closed PIC.

Fix codeGeneratorToComputeAccessorDepth.  I
went to all that trouble in VMMaker.oscog-eem.873
to not rely on cogitClass et al but forgot to change
the operative case.