CogVM source as per VMMaker.oscog-eem.886
Fix regression in V# become youngReferrers management in VMMaker.oscog-eem.882.
We /must/ prune young referrers if mapObjectReferencesInMachineCodeForBecome
removes a cog method from youngReferrers because it may get added back and
youngReferrers cannot contain duplicates.

Commit this separate from VMMaker.oscog-eem.885 to have it separate from the
flush caches on Spur become changes from VMMaker.oscog-eem.883/884.

To be clear, this is a branch.  The head is 885.


git-svn-id: http://squeakvm.org/svn/squeak/branches/Cog@3083 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: c5ca2064348c42af81d9d3ab7b17eceef647945a