Install the callPrimitiveBytecode in the Interpreter's bytecodeDispatchTable on Spur.
Modify callPrimitiveBytecode to not invoke
unknownBytecode processing if at the first bytecode
of a primitive method.

Correct sign comparison of instructionPointer in
justActivateNewMethod.

Relax the validInstructionPointer:inMethod:framePointer:
assert to accept any pc in initialPC to self size range
now that callPrimitiveBytecode is more forgiving.

Speed up primitiveMarkUnwindMethod &
primitiveMarkHandlerMethod in the StackInterpreter by
setting them to 0 in the primitive table.