CogVM source as per VMMaker.oscog-eem.887
Install the callPrimitiveBytecode in the Interpreter's
bytecodeDispatchTable on Spur.

Modify callPrimitiveBytecode to not invoke unknownBytecode processing if at the
first bytecode of a primitive method.

Correct sign comparison of instructionPointer in justActivateNewMethod.

Relax the validInstructionPointer:inMethod:framePointer:
assert to accept any pc in initialPC to self size range
now that callPrimitiveBytecode is more forgiving.

Speed up primitiveMarkUnwindMethod & primitiveMarkHandlerMethod in the
StackInterpreter by setting them to 0 in the primitive table.


git-svn-id: http://squeakvm.org/svn/squeak/branches/Cog@3085 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: c6deb4e8c0723a25c9091205f421a504fe829794