VMMaker.oscog-eem.890 Spur: Fix one-way become for classes with and without copyHash, primarily by fixing allInstances.  One-way become for classes causes duplicates in the class table (either that or an allInstances scan would be needed as part of the become to change instances referring to the deleted class index, which would be very slow).  So allInstances must be able to cope with duplicates.  Hence split it into a fast path common case when the class in question is not duplicated, and a slower path when it is.  Make both the marking phase of GC and allInstances check for and eliminate refereces to duplicate entries at wrong/obsolete class indices.
Fix markAndShouldScan: to not scan the pun classes
of non-objects on the heap such as implicit receiver
caches and Sista counters.

Eliminate the classTableBitmap premature
optimization.  All the information we need is in the
cassTable and the class's hashes therein.

Document the failure return of
allocateSlotsInOldSpace:bytes:format:classIndex:

Change pinObject: to answer 0 on failure and the
(possibly moved) object on success.  Much easier
than having to check and follow forwarding pointer.

Simulator:
Eliminate the halt in Spur globalGarbageCollect,
the lemming approach raising an error in the copy
being much more useful.

Fix FakeStdinStream and FilePluginSimulator>>
sqFile:Read:Into:At: for buffer sizes > 1.

misc:
fix some C compilation warnings