CogVM source as per VMMaker.oscog-eem.890
Spur:
Fix one-way become for classes with and without copyHash, primarily by fixing
allInstances.  One-way become for classes causes duplicates in the class table
(either that or an allInstances scan would be needed as part of the become to
change instances referring to the deleted class index, which would be slow).
So allInstances must be able to cope with duplicates.  Hence split it into a
fast path common case when the class in question is not duplicated, and a slower
path when it is.  Make both the marking phase of GC and allInstances check for
and eliminate refereces to duplicate entries at wrong/obsolete class indices.

Fix markAndShouldScan: to not scan the pun classes of non-objects on the heap
such as implicit receiver caches and Sista counters.

Eliminate the classTableBitmap premature optimization.  All the information we
need is in the cassTable and the class's hashes therein.

Change pinObject: to answer 0 on failure and the (possibly moved) object on
success.  Much easier than having to check and follow forwarding pointer.
The changes to platforms/Cross/plugins/FilePlugin/sqFilePluginBasicPrims.c
in r3092 in a Spur MT VM depend on this.

Fix some C compilation warnings


git-svn-id: http://squeakvm.org/svn/squeak/branches/Cog@3093 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: f914a657d7395997e850465b4abc73ccdf230821