CogVM source as per VMMaker.oscog-eem.913
Fix baaaad bug in checking for still-married contexts in Spur.  One *cannot*
simply follow what is assumed to be the frameContext field of what is presumed
to be a context.  On Spur we need to *know* whether we have a valid frameContext
field, and for that we have to know there is a valid frame.  So refactor, moving
isFrame:onPage: from StackInterpreterPrimitives to StackInterpreter, and use it
to validate the frame pointer of a maybe married context before testing the
context for being forwarded.


git-svn-id: http://squeakvm.org/svn/squeak/branches/Cog@3111 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: 53405775ad0abfc8bbb080cb89e7c5fc93fc4308