Spur: Implement safe access to contexts and methods when creating image segments (since internally married contexts and jitted methods have hidden state encoded in odd ways).
As a result:
Fix baaad bug in SpurMemoryManager>>lastPointerOf:
for contexts, caused by transcription from
(New)ObjectMemory where the method works on a
wing and a prayer because
	BaseHeaderSize = BytesPerOop.
And I can't be arsed to fix it.  It's, cough, grody code.

Refactor decoding of context instruction pointers so
there is only one copy of the code.

Fix baaad bug in accessibleObjectAfter: and add an
assert to primitiveNextObject to catch similar bugs.
Can now enumerate objects in new space, which will
sort-of work, but code should really use allObjects.
Will fix at the image level as time allows.

Eliminate isInSurvivorSpace: in favour of the identical
isInPastSpace:.

All:
Fix the comments in loadImageSegmentFrom:outPointers:;
the array is not truncated to zero length; it retains
its version info.

Fix storeShort16:ofObject:withValue: to accept
negative values.